+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Release Notes for command-line
 
+## Unreleased
+- Added `command-line/fields/create` for creating new custom fields of any type.
+
 ## 1.0.0
 - Initial release